+++ src/util/api.mjs
@@ -1,8 +1,11 @@
 
 class Api {
     get allBonusTypes() { return Object.values(this.bonusTypeMap); }
+    get allBonusTypesKeys() { return Object.keys(this.bonusTypeMap); }
     get allTargetTypes() { return Object.values(this.targetTypeMap); }
+    get allTargetTypesKeys() { return Object.keys(this.targetTypeMap); }
     get allGlobalTypes() { return Object.values(this.globalTypeMap); }
+    get allGlobalTypesKeys() { return Object.keys(this.globalTypeMap); }
     bonusTypeMap = {};
     globalTypeMap = {};
     sources = {};